feat: implement dynamic pre-context phase
This pre-context phase will not be part of the YAML that the workflow
defines and will be automatically provided by Rover depending on the
context. For example, if we are iterating, this step will be injected
so that the information about previous iteration is known to the
agent.

The `rover-agent` allows for a generic `--pre-context-file` that
points to this pre-context information that will be injected.